Restructure project directory and packaging
