+++ main.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"log"
 	"os"
+
+	"github.com/tsoud/GoTGit.git/cmd"
 )
 
 func main() {
@@ -13,12 +15,12 @@ func main() {
 
 	switch command := os.Args[1]; command {
 	case "init":
-		initCmdArgs, quiet := setupInitCommand()
+		initCmdArgs, quiet := cmd.SetupInitCommand()
 		initCmdArgs.Parse(os.Args[2:])
-		initCmdHandler(quiet)
+		cmd.InitCmdHandler(quiet)
 
 	case "cat-file":
-		catFileCmdArgs, pprint := setupCatFileCmd()
+		catFileCmdArgs, pprint := cmd.SetupCatFileCmd()
 		catFileCmdArgs.Parse(os.Args[2:])
 		files := catFileCmdArgs.Args()
 		if len(files) < 1 {
@@ -27,10 +29,10 @@ func main() {
 		if len(files) > 1 {
 			log.Printf("More than one file given.\nOnly showing 1st file %s:", files[0])
 		}
-		catFileCmdHandler(pprint, files[0])
+		cmd.CatFileCmdHandler(pprint, files[0])
 
 	case "hash-object":
-		hashObjCmdArgs, write, objType := setupHashObjectCmd()
+		hashObjCmdArgs, write, objType := cmd.SetupHashObjectCmd()
 		hashObjCmdArgs.Parse(os.Args[2:])
 		file := hashObjCmdArgs.Args()
 		if len(file) < 1 {
@@ -39,7 +41,7 @@ func main() {
 		if len(file) > 1 {
 			log.Printf("Found more than one file. Only hashing %s\n", file[0])
 		}
-		hashObjectCmdHandler(write, objType, file[0])
+		cmd.HashObjectCmdHandler(write, objType, file[0])
 
 	default:
 		fmt.Fprintf(os.Stderr, "Unknown command %s\n", command)
